user_dashboard_api: extract customer info and user logic to api layer
